Project 구조 개선
