+++ server/app/admin/src/main/kotlin/com/starter/admin/AdminApplication.kt
@@ -1,21 +1,23 @@
 package com.starter.admin
 
+import com.starter.core.clients.internal.InternalClientConfig
 import com.starter.core.common.config.CommonConfig
 import com.starter.core.jasypt.config.JasyptConfig
 import com.starter.core.rdb.config.DataSourceConfig
 import com.starter.core.security.config.SecurityConfig
+import com.starter.core.tempfile.config.TempFileConfig
 import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan
 import org.springframework.boot.runApplication
 import org.springframework.context.annotation.Import
 
-@ConfigurationPropertiesScan
 @SpringBootApplication
 @Import(
     CommonConfig::class,
     JasyptConfig::class,
     SecurityConfig::class,
-    DataSourceConfig::class
+    DataSourceConfig::class,
+    InternalClientConfig::class,
+    TempFileConfig::class,
 )
 class AdminApplication
