+++ server/app/admin/src/main/kotlin/com/starter/admin/app/user/model/UserCreateRequest.kt
@@ -1,0 +1,6 @@
+package com.starter.admin.app.user.model
+
+data class UserCreateRequest(
+    val email: String,
+    val userName: String,
+)
