+++ server/app/admin/src/main/kotlin/com/starter/admin/config/WebMvcConfig.kt
@@ -7,7 +7,6 @@ import org.springframework.context.annotation.Configuration
 import org.springframework.format.FormatterRegistry
 import org.springframework.format.datetime.standard.DateTimeFormatterRegistrar
 import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
 import java.time.format.DateTimeFormatter
 
@@ -15,11 +14,6 @@ import java.time.format.DateTimeFormatter
 class WebMvcConfig(
     private val objectMapper: ObjectMapper,
 ) : WebMvcConfigurer {
-
-    override fun addResourceHandlers(registry: ResourceHandlerRegistry) {
-        registry.addResourceHandler("/static/**")
-            .addResourceLocations("classpath:/static/")
-    }
 
     @Bean
     fun mappingJackson2HttpMessageConverter(): MappingJackson2HttpMessageConverter {
