--- server/app/db-migration/build.gradle.kts
@@ -1,24 +1,0 @@
-tasks {
-    application {
-        mainClass.set("com.starter.dbmigration.DbMigrationApplication")
-    }
-
-    bootJar {
-        enabled = true
-    }
-
-    jar {
-        enabled = false
-    }
-}
-
-dependencies {
-    val flywayVersion = "9.21.1"
-
-    implementation(project(":server:core"))
-    testImplementation(project(":server:core"))
-
-    // flyway
-    implementation("org.flywaydb:flyway-core:$flywayVersion")
-    implementation("org.flywaydb:flyway-mysql:$flywayVersion")
-}
