--- server/app/db-migration/src/main/resources/application-dev.yml
@@ -1,9 +1,0 @@
-logging:
-  level:
-    root: DEBUG
-    org.springframeworkboot.autoconfigure: ERROR
-
-spring:
-  config:
-    activate:
-      on-profile: dev
