--- server/app/db-migration/src/main/resources/application-local.yml
@@ -1,12 +1,0 @@
-logging:
-  level:
-    root: INFO
-    com.starter: DEBUG
-    org:hibernate.SQL: debug
-    type.descriptor.sql: TRACE
-
-
-spring:
-  config:
-    activate:
-      on-profile: local
