+++ server/app/file/src/main/kotlin/com/starter/file/FileApplication.kt
@@ -5,6 +5,7 @@ import com.starter.core.jasypt.config.JasyptConfig
 import com.starter.core.rdb.config.DataSourceConfig
 import com.starter.core.s3.config.S3Config
 import com.starter.core.security.config.SecurityConfig
+import com.starter.core.tempfile.config.TempFileConfig
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan
 import org.springframework.boot.runApplication
@@ -18,6 +19,7 @@ import org.springframework.context.annotation.Import
     SecurityConfig::class,
     DataSourceConfig::class,
     S3Config::class,
+    TempFileConfig::class,
 ])
 class FileApplication
