+++ server/app/file/src/main/resources/application-local.yml
@@ -6,6 +6,20 @@ logging:
       SQL: debug
       type.descriptor.sql: trace
 
-
 file:
   upload-path: file-dir/upload/
+
+spring:
+  config:
+    activate:
+      on-profile: local
+  flyway:
+    locations: flyway/migrations
+    table: file_flyway_schema_history
+    enabled: true
+    url: jdbc:mysql://localhost:3306/starter
+    encoding: UTF-8
+    user: starter
+    password: starter
+    baseline-version: 0
+    baseline-on-migrate: true
