--- server/app/user/build.gradle.kts
@@ -1,27 +1,0 @@
-tasks {
-    application {
-        mainClass.set("com.starter.user.UserApplicationKt")
-    }
-
-    bootJar {
-        enabled = true
-    }
-
-    jar {
-        enabled = false
-    }
-}
-
-dependencies {
-    val jwtVersion = "0.11.5"
-
-    implementation(project(":server:core"))
-    testImplementation(project(":server:core"))
-
-    implementation("io.jsonwebtoken:jjwt-api:$jwtVersion")
-    implementation("io.jsonwebtoken:jjwt-impl:$jwtVersion")
-    implementation("io.jsonwebtoken:jjwt-jackson:$jwtVersion")
-
-    testImplementation("org.springframework.boot:spring-boot-starter-test")
-
-}
