--- server/app/user/src/main/kotlin/com/starter/user/app/user/model/UserCreateRequest.kt
@@ -1,6 +1,0 @@
-package com.starter.user.app.user.model
-
-data class UserCreateRequest(
-    val email: String,
-    val userName: String,
-)
