--- server/app/user/src/main/kotlin/com/starter/user/config/SpringDocConfig.kt
@@ -1,25 +1,0 @@
-package com.starter.user.config
-
-import io.swagger.v3.oas.models.OpenAPI
-import io.swagger.v3.oas.models.info.Info
-import io.swagger.v3.oas.models.servers.Server
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-
-@Configuration
-class SpringDocConfig {
-
-    @Bean
-    fun openAPI(): OpenAPI {
-        return OpenAPI()
-            .info(
-                Info()
-                    .title("user API document")
-                    .version("v1.0.0")
-            )
-            .addServersItem(
-                Server()
-                    .url("/")
-            )
-    }
-}
