--- server/app/user/src/main/resources/application-dev.yml
@@ -1,12 +1,0 @@
-logging:
-  level:
-    root: DEBUG
-    org:
-      springframework:
-        boot:
-          autoconfigure: ERROR
-
-spring:
-  config:
-    activate:
-      on-profile: dev
