--- server/app/user/src/main/resources/application-local.yml
@@ -1,13 +1,0 @@
-logging:
-  level:
-    root: INFO
-    com.starter: DEBUG
-    org:hibernate:
-      SQL: debug
-      type.descriptor.sql: trace
-
-
-spring:
-  config:
-    activate:
-      on-profile: local
