--- server/app/user/src/main/resources/application-prod.yml
@@ -1,16 +1,0 @@
-logging:
-  level:
-    root: INFO
-    org:
-      springframework:
-        boot:
-          autoconfigure: ERROR
-
-spring:
-  config:
-    activate:
-      on-profile: prod
-
-springdoc:
-  api-docs:
-    enabled: false
