--- server/core/src/main/kotlin/com/starter/core/tempfile/config/TempFIleConfig.kt
@@ -1,12 +1,0 @@
-package com.starter.core.tempfile.config
-
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan
-import org.springframework.context.annotation.ComponentScan
-import org.springframework.context.annotation.Configuration
-import org.springframework.scheduling.annotation.EnableScheduling
-
-@Configuration
-@ComponentScan("com.starter.core.tempfile")
-@ConfigurationPropertiesScan("com.starter.core.tempfile")
-@EnableScheduling
-class TempFIleConfig
