+++ server/core/src/main/kotlin/com/starter/core/tempfile/scheduler/TempFileSystemScheduler.kt
@@ -1,16 +1,16 @@
 package com.starter.core.tempfile.scheduler
 
-import com.starter.core.tempfile.service.TempFileSystemProcessor
+import com.starter.core.tempfile.service.TempFileSystemRotationService
 import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.stereotype.Component
 
 @Component
 class TempFileSystemScheduler(
-    private val tempFileSystemProcessor: TempFileSystemProcessor,
+    private val tempFileSystemRotationService: TempFileSystemRotationService,
 ) {
 
     @Scheduled(cron ="0 0 */1 * * *")
     fun hourlyDeleteLocalFiles() {
-        tempFileSystemProcessor.deleteDir()
+        tempFileSystemRotationService.deleteDir()
     }
 }
